Eliminar Productos y Update Persistencia

+++ Contador de Cuotas/Models.swift
@@ -11,10 +11,10 @@ class Producto: Identifiable, Codable {
     var id = UUID()
     var nombre: String
     var cuotasRestantes: Int
-    var precioPorCuota: Double
+    var precioPorCuota: Int
     var nombreTarjeta: String // Nuevo campo
 
-    init(nombre: String, cuotasRestantes: Int, precioPorCuota: Double, nombreTarjeta: String) {
+    init(nombre: String, cuotasRestantes: Int, precioPorCuota: Int, nombreTarjeta: String) {
         self.nombre = nombre
         self.cuotasRestantes = cuotasRestantes
         self.precioPorCuota = precioPorCuota
